fix(kda): remove TORCH_MODE from prepare kernel
